experiments/2026/15-4-2026/requirements.txt
torch>=2.0
transformers>=4.40
datasets
scikit-learn
scipy
numpy
matplotlib
tqdm
python-dotenv
inspect-ai
bitsandbytes
accelerate